Compilation fix
git-svn-id: file:///home/roberto/Git/svn@6520 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 8fce51f50456df59312c46b683703ad6ac21b3a5

+++ trunk/xmipp/libraries/data/image.h
@@ -114,7 +114,8 @@ struct ImageFHandler
 
 
 /// Returns memory size of datatype
-unsigned long gettypesize(DataType type);
+size_t gettypesize(DataType type);
+
 /** Convert datatype string to datatypr enun */
 int datatypeString2Int(std::string s);
